+++ src/components/form/basicControls/textArea/types.ts
@@ -1,15 +1,14 @@
 import type { Slot } from 'vue'
-import type { IBaseFieldProps, InputMode } from '../BaseField'
+import type { IBaseFieldProps, IBaseMessage, InputMode } from '../BaseField'
 
 export const TEXTAREA_RESIZE_OPTIONS = ['none', 'vertical', 'horizontal', 'both'] as const
 
 export type TextAreaResize = (typeof TEXTAREA_RESIZE_OPTIONS)[number]
 
-export interface UiTextAreaProps extends IBaseFieldProps {
+export interface UiTextAreaProps extends Omit<IBaseFieldProps, 'errorMessages' | 'infoMessage'> {
   autocomplete?: string
   invalid?: boolean
   maxlength?: number
-  message?: string
   rows?: number
   resize?: TextAreaResize
   inputMode?: InputMode
@@ -24,6 +23,5 @@ export interface UiTextAreaEmits {
 
 export interface UiTextAreaSlots {
   label?: Slot
-  message?: Slot
-  errorMessages?: Slot
+  message?: Slot<{ message: IBaseMessage }>
 }